requirements/base.txt
Django>=5.2,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.5,<6.0
python-decouple>=3.8,<4.0
django-redis>=5.4,<6.0
redis>=5.2,<6.0
django-ratelimit>=4.1,<5.0
Pillow>=11.0,<12.0
dj-database-url>=2.3,<3.0
psycopg[binary]>=3.2,<4.0
drf-spectacular>=0.28,<1.0
httpx>=0.28,<1.0
channels>=4.2,<5.0
channels-redis>=4.2,<5.0
celery>=5.4,<6.0
flower>=2.0,<3.0
daphne>=4.1,<5.0
